feat: create common article info component
