+++ src/templates/CommonPostTemplate.tsx
@@ -7,8 +7,6 @@ import { BLOCKS, MARKS, INLINES } from '@contentful/rich-text-types';
 import { makeStyles, useMediaQuery } from '@material-ui/core';
 import Typography from '@material-ui/core/Typography';
 
-import DateRangeIcon from '@material-ui/icons/DateRange';
-
 import Layout from '@/components/others/Layout';
 import CommonTemplateH1 from '@/components/basic/title/CommonTemplateH1';
 import CommonTemplateH2 from '@/components/basic/title/CommonTemplateH2';
@@ -19,60 +17,17 @@ import CommonTemplateQuote from '@/components/basic/paragraph/CommonTemplateQuot
 import CommonTemplateUl from '@/components/basic/list/CommonTemplateUl';
 import CommonTemplateOl from '@/components/basic/list/CommonTemplateOl';
 import CommonEntryCard from '@/components/parts/CommonEntryCard';
+import CommonArticleInfo from '@/components/parts/CommonArticleInfo';
 
 const useStyles = makeStyles({
     eyeCatchStyles: {
         width: '100%',
         height: '24vh',
     },
-    articleInfoBox: {
-        width: '90%',
-        maxWidth: '1080px',
-        margin: '3vw auto',
-    },
-    articleTitle: {
-        margin: '.5em 0'
-    },
-    articleCreatedDate: {
-        display: 'flex',
-        alignItems: 'center',
-        margin: '.5em 0'
-    },
-    dateIcon: {
-        color: '#ffffff',
-    },
     articleContentBox: {
         width: '90%',
         maxWidth: '1080px',
         margin: '2em auto'
-    },
-    articleContentUl: {
-        color: '#ffffff',
-    },
-    articleContentOl: {
-        color: '#ffffff',
-    },
-});
-
-const useSPStyles = makeStyles({
-    dateText: {
-        fontSize: '.8em'
-    },
-    dateIcon: {
-        marginRight: '.25em',
-        width: '.8em',
-        height: '.8em',
-    },
-});
-
-const usePCStyles = makeStyles({
-    dateText: {
-        fontSize: '1em'
-    },
-    dateIcon: {
-        marginRight: '.75em',
-        width: '1em',
-        height: '1em',
     },
 });
 
@@ -113,7 +68,6 @@ const CommonPostTemplate = ({ pageContext }) => {
     const eyeCatch = getImage(post.eyeCatch);
     const inXs = useMediaQuery('(max-width: 600px)');
     const classes = useStyles();
-    const dividedClasses = inXs ? useSPStyles() : usePCStyles();
 
     return (
         <Layout pathname=''>
@@ -123,13 +77,10 @@ const CommonPostTemplate = ({ pageContext }) => {
                     alt={`${post.title} eyeCatch`}
                     className={classes.eyeCatchStyles}
                 />
-                <div className={classes.articleInfoBox}>
-                    <Typography variant='h1'>{post.title}</Typography>
-                    <div className={classes.articleCreatedDate}>
-                        <DateRangeIcon className={`${classes.dateIcon} ${dividedClasses.dateIcon}`} />
-                        <Typography className={dividedClasses.dateText}>{post.createdAt}</Typography>
-                    </div>
-                </div>
+                <CommonArticleInfo
+                    title={post.title}
+                    createdAt={post.createdAt}
+                />
                 <section className={classes.articleContentBox}>
                     {post.content && renderRichText(post.content, options)}
                 </section>
